requirements.txt
# Core dependencies
python-dotenv>=1.0.0
python-dateutil>=2.8.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Google Sheets API
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
kaleido>=0.2.1

# Add your project dependencies below
# requests>=2.31.0
# beautifulsoup4>=4.12.0
